+++ src/main/java/com/cheeup/service/member/MemberService.java
@@ -1,10 +1,10 @@
 package com.cheeup.service.member;
 
-import com.cheeup.web.dto.ReadMemberDto;
-import com.cheeup.web.dto.UpdateMemberDto;
+import com.cheeup.web.dto.member.ReadMemberDto;
+import com.cheeup.web.dto.member.UpdateMemberDto;
 
 public interface MemberService {
-    ReadMemberDto.Response getMemberInfo(long id);
+    ReadMemberDto.ResponseDto getMemberInfo(long id);
 
     void updateMemberInfo(long id, UpdateMemberDto.Request request);
 }
